feat: added possibility to create a progress json file for CLI training processes
Added two new parameters to CLI commands:
--save-status enable the feature of saving the status as a json file while running the training
--status-filename let you set the json file name

To easily connect brush to a blender plugin we needed a simple way to get the progress from the brush process. After various failed tries on Blender/Python we had to find another solution.
With this method we can retrieve the current process just accessing a json file.